contours/requirements.txt
opencv-python
matplotlib
numpy
scipy
argparse
Pillow
